Unify events controllers for scopes and issues
